+++ tests/server/blueprints/cases/test_cases_views.py
@@ -2,7 +2,7 @@
 import datetime
 
 from bson.objectid import ObjectId
-from flask import current_app, json, url_for
+from flask import json, url_for
 
 from scout.constants import CUSTOM_CASE_REPORTS
 from scout.server.blueprints.cases.views import parse_raw_gene_ids, parse_raw_gene_symbols
@@ -416,6 +416,30 @@ def test_case_sma(app, case_obj, institute_obj):
                 "cases.sma",
                 institute_id=institute_obj["internal_id"],
                 case_name=case_obj["display_name"],
+            )
+        )
+
+        # THEN it should return a page
+        assert resp.status_code == 200
+
+
+def test_case_fusion(app, adapter, fusion_case_obj, institute_obj):
+    """Test the RNA fusion case page."""
+
+    # GIVEN an initialized app
+    with app.test_client() as client:
+        # GIVEN a valid user, case and institute
+        client.get(url_for("auto_login"))
+
+        # GIVEN a database containing a fusion case object
+        assert store.case_collection.insert_one(fusion_case_obj)
+
+        # WHEN accessing the RNA fusion case
+        resp = client.get(
+            url_for(
+                "cases.case",
+                institute_id=institute_obj["internal_id"],
+                case_name=fusion_case_obj["display_name"],
             )
         )